+++ .github/workflows/verify-build.yml
@@ -55,8 +55,8 @@ jobs:
           go-version-file: go.mod
       - run: make unit-tests
         name: Run Unit Tests
-  integration-tests:
-    name: Integration Tests
+  integration-basic:
+    name: Integration Basic Tests
     needs: basic-checks
     runs-on: ${{ github.repository == 'hyperledger/fabric-x-orderer' && 'ubuntu-24.04' }}
     steps:
@@ -68,8 +68,38 @@ jobs:
         name: Install Go
         with:
           go-version-file: go.mod
-      - run: make integration-tests
-        name: Run Integration Tests
+      - run: make integration-basic
+        name: Run Integration Basic Tests
+  integration-faulttolerance:
+    name: Integration Fault Tolerance Tests
+    needs: basic-checks
+    runs-on: ${{ github.repository == 'hyperledger/fabric-x-orderer' && 'ubuntu-24.04' }}
+    steps:
+      - uses: actions/checkout@v5
+        with:
+          fetch-depth: 0
+        name: Checkout Fabric-X Code
+      - uses: actions/setup-go@v6
+        name: Install Go
+        with:
+          go-version-file: go.mod
+      - run: make integration-faulttolerance
+        name: Run Integration Fault Tolerance Tests
+  integration-reconfig:
+    name: Integration Reconfig Tests
+    needs: basic-checks
+    runs-on: ${{ github.repository == 'hyperledger/fabric-x-orderer' && 'ubuntu-24.04' }}
+    steps:
+      - uses: actions/checkout@v5
+        with:
+          fetch-depth: 0
+        name: Checkout Fabric-X Code
+      - uses: actions/setup-go@v6
+        name: Install Go
+        with:
+          go-version-file: go.mod
+      - run: make integration-reconfig
+        name: Run Integration Reconfig Tests
   sample-tests:
     name: Sample Tests
     needs: basic-checks
